Refactor users app
- Move Chart class to utils
- Move relevant functions to utils
- Move counselor_view and my_counselees functions to counselors
- Move student_view function to students
