+++ users/views/users.py
@@ -1,17 +1,13 @@
-import plotly.graph_objects as go
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.forms import AuthenticationForm
-from django.db.models import Count
 from django.http import HttpResponseRedirect
-from django.http.response import JsonResponse
 from django.shortcuts import render
 from django.urls import reverse
 from django.views.generic import TemplateView, FormView
-from nltk.metrics.distance import edit_distance
 
-from home import views as home_views
-from home.models import Achievement, Organization
-from users.models import Counselor, Student, User
+from users.models import User
+from users.views.counselors import counselor_view
+from users.views.students import student_view
 
 
 class SignUpView(TemplateView):
@@ -52,164 +48,3 @@ class LoginView(FormView):
             return HttpResponseRedirect(reverse("index"))
         else:
             return self.form_invalid(form)
-
-
-def counselor_view(request, user_obj):
-    achievements = Achievement.objects.all()
-    counselor = Counselor.objects.get(user=user_obj)
-    context = {
-        'id': counselor.id,
-        'user_type': Counselor.type,
-    }
-
-    if request.method == 'POST':
-        usn = request.POST.get('usn', '').upper()
-        year = request.POST.get('year')
-        achievement_type = request.POST.get('achievement_type', '').title()
-        organization = request.POST.get('organization')
-        order = request.POST.get('sort_by')
-        counselees = request.POST.get('my_counselees')
-
-        if usn:
-            student = Student.objects.filter(usn=usn).first()
-            achievements = achievements.filter(holders=student) if student else []
-        elif counselees:
-            students = Student.objects.filter(counselor=counselor)
-            achievements = achievements.filter(holders__in=students) if students else []
-
-        if achievement_type:
-            achievements = achievements.filter(type=achievement_type)
-            if not achievements.exists():
-                context['type_suggestions'] = get_suggestions('type', achievement_type)
-
-        achievements = achievements.filter(academic_year=year) if year else achievements
-
-        if organization:
-            org_obj = Organization.objects.filter(name=organization).first()
-            achievements = achievements.filter(organization=org_obj)
-            if not achievements.exists():
-                context['org_suggestions'] = get_suggestions('organization', organization)
-
-        achievements = achievements.order_by(order) if order else achievements
-
-        context.update({'usn': usn, 'year': year, 'achievement_type': achievement_type, 'org': organization})
-
-    chart = Chart.achievement_count_per_type()
-    context.update({'achievements': achievements, "achievements_links": zip_links(achievements), "fig": chart})
-
-    return render(request, "users/counselor_view.html", context)
-
-
-def my_counselees(request):
-    if not request.user.is_authenticated:
-        return render(request, "users/logout.html", {'message': 'You must be signed in.'})
-    if not User.objects.get(username=request.user.username).is_counselor:
-        return render(request, "users/my_counselees.html",
-                      {'message': 'Only Counselors are Authorized to this page.'})
-
-    me = Counselor.objects.get(user=request.user)
-    students = Student.objects.filter(counselor=me).order_by('usn')
-    context = {
-        'id': me.id,
-        'students': students,
-    }
-    return render(request, "users/my_counselees.html", context)
-
-
-def student_view(request, user_obj):
-    if request.method == 'POST':
-        return home_views.edit_achievement(request)
-
-    student = Student.objects.get(user=user_obj)
-    achievements = student.achievements.all()
-
-    context = {
-        'usn': student.usn,
-        'counselor': student.counselor.id,
-        'user_type': Student.type,
-        'achievements': achievements,
-        'achievements_links': zip_links(achievements),
-    }
-    return render(request, "users/student_view.html", context)
-
-
-def get_suggestions(key, word):
-    corpus = []
-    if key == 'type':
-        corpus = list(Achievement.objects.values_list('type', flat=True).distinct())
-    elif key == 'organization':
-        corpus = list(Organization.objects.values_list('name', flat=True).distinct())
-
-    mn = 10
-    distances = []
-    for cor in corpus:
-        ed = edit_distance(word, cor)
-        mn = min(mn, ed)
-        distances.append(ed)
-
-    suggestions = []
-    cor_ed = zip(corpus, distances)
-    for cor, ed in cor_ed:
-        if ed == mn:
-            suggestions.append(cor)
-
-    return suggestions
-
-
-def zip_links(achievements):
-    links = [f"download?achid={achievement.id}" for achievement in achievements]
-    return zip(achievements, links)
-
-
-def usn_complete(request):
-    if 'term' in request.GET:
-        qs = Student.objects.filter(usn__icontains=request.GET.get('term')).order_by('usn')
-        usn_list = list(qs.values_list('usn', flat=True))
-        return JsonResponse(usn_list, safe=False)
-    return JsonResponse(None)
-
-
-def achievement_type_complete(request):
-    if 'term' in request.GET:
-        qs = Achievement.objects.filter(type__icontains=request.GET.get('term'))
-        type_list = list(qs.values_list('type', flat=True).distinct())
-        return JsonResponse(type_list, safe=False)
-    return JsonResponse(None)
-
-
-def organization_type_complete(request):
-    if 'term' in request.GET:
-        qs = Organization.objects.filter(type__icontains=request.GET.get('term'))
-        type_list = list(qs.values_list('type', flat=True).distinct())
-        return JsonResponse(type_list, safe=False)
-    return JsonResponse(None)
-
-
-class Chart:
-
-    @staticmethod
-    def achievement_count_per_student():
-        students = Student.objects.all().order_by("usn")
-        labels = [str(student) for student in students]
-        values = [student.achievements.count() for student in students]
-        chart = go.Figure(data=[go.Pie(labels=labels, values=values, hole=0.3)])
-        return chart.to_html()
-
-    @staticmethod
-    def achievement_count_per_year():
-        achievements = Achievement.objects.values("academic_year").annotate(count=Count('academic_year'))
-        labels = [achievement["academic_year"] for achievement in achievements]
-        values = [achievement["count"] for achievement in achievements]
-        chart = go.Figure(data=[go.Bar(x=labels, y=values)])
-        chart.update_yaxes(title_text="Achievements Count", dtick=1)
-        chart.update_xaxes(title_text="Academic Year", range=[2010, 2022], dtick=1)
-        return chart.to_html()
-
-    @staticmethod
-    def achievement_count_per_type():
-        achievements = Achievement.objects.values("type").annotate(count=Count("type"))
-        labels = [achievement["type"] for achievement in achievements]
-        values = [achievement["count"] for achievement in achievements]
-        chart = go.Figure(data=[go.Pie(labels=labels, values=values, title="Achievements Types")])
-        chart.update_traces(textposition='inside', textinfo='percent+label')
-        return chart.to_html()
